+++ src/scribbler/__tests__/Line.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, test } from "@jest/globals";
 import ansiEscapesSerializer from "jest-serializer-ansi-escapes";
-import { Color } from "../../Color.js";
-import { Scribbler } from "../../Scribbler.js";
+import { Color } from "../Color.js";
+import { Scribbler } from "../Scribbler.js";
 import { Line } from "../Line.js";
 
 expect.addSnapshotSerializer(ansiEscapesSerializer);
