+++ src/scribbler/__tests__/Text.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect, test } from "@jest/globals";
 import ansiEscapesSerializer from "jest-serializer-ansi-escapes";
-import { Color } from "../../Color.js";
-import { Scribbler } from "../../Scribbler.js";
+import { Color } from "../Color.js";
+import { Scribbler } from "../Scribbler.js";
 import { Text } from "../Text.js";
 
 expect.addSnapshotSerializer(ansiEscapesSerializer);
