+++ src/scribbler/components/describeNameText.tsx
@@ -1,5 +1,5 @@
+import { Line } from "../Line.js";
 import { Scribbler } from "../Scribbler.js";
-import { Line } from "./Line.js";
 
 export function describeNameText(name: string, indent = 0): JSX.Element {
   return <Line indent={indent + 1}>{name}</Line>;
