+++ src/scribbler/components/fileStatusText.tsx
@@ -2,10 +2,10 @@ import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { type FileResultStatus, ResultStatus } from "#result";
 import { Color } from "../Color.js";
+import { Line } from "../Line.js";
 import { Scribbler } from "../Scribbler.js";
-import { Line } from "./Line.js";
+import { Text } from "../Text.js";
 import { RelativePathText } from "./RelativePathText.js";
-import { Text } from "./Text.js";
 
 class FileNameText implements JSX.ElementClass {
   constructor(readonly props: { filePath: string }) {}
