+++ src/scribbler/components/testNameText.tsx
@@ -1,7 +1,7 @@
 import { Color } from "../Color.js";
+import { Line } from "../Line.js";
 import { Scribbler } from "../Scribbler.js";
-import { Line } from "./Line.js";
-import { Text } from "./Text.js";
+import { Text } from "../Text.js";
 
 class StatusText implements JSX.ElementClass {
   constructor(readonly props: { status: "fail" | "pass" | "skip" | "todo" }) {}
