+++ src/scribbler/index.ts
@@ -5,9 +5,7 @@ export { fileStatusText } from "./components/fileStatusText.js";
 export { fileViewText } from "./components/fileViewText.js";
 export { formattedText } from "./components/formattedText.js";
 export { helpText } from "./components/helpText.js";
-export { Line } from "./components/Line.js";
 export { summaryText } from "./components/summaryText.js";
 export { testNameText } from "./components/testNameText.js";
-export { Text } from "./components/Text.js";
 export { usesCompilerStepText } from "./components/usesCompilerStepText.js";
 export { Scribbler, type ScribblerOptions } from "./Scribbler.js";
